dom/tests/mochitest/ajax/offline/test_redirectManifest.html

<html xmlns="http://www.w3.org/1999/xhtml" manifest="https://example.com/tests/dom/tests/mochitest/ajax/offline/manifestRedirect.sjs">
<head>
<title>Fail update on manifest redirection test</title>

<script src="/tests/SimpleTest/SimpleTest.js"></script>
<script type="text/javascript" src="/tests/dom/tests/mochitest/ajax/offline/offlineTests.js"></script>
<link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />

<script class="testbody" type="text/javascript">

/**
 */

function manifestCached()
{
  OfflineTests.ok(false, "Manifest must not be cached");
  finish();
}

function manifestError()
{
  OfflineTest.ok(true, "Error expected");
  finish();
}

function finish()
{
  OfflineTest.teardownAndFinish();
}

SimpleTest.waitForExplicitFinish();

if (OfflineTest.setup()) {
  applicationCache.onerror = OfflineTest.priv(manifestError);
  applicationCache.oncached = OfflineTest.priv(manifestCached);
}

</script>

</head>

<body>
</body>
</html>
